PARKHOMENKO_BOT/requirements.txt
# Telegram Bot Framework
aiogram==3.7.0

# Environment variables
python-dotenv==1.0.0

# Database
aiosqlite==0.19.0

# HTTP requests
requests==2.31.0
aiohttp==3.11.11
httpx==0.28.1

# RouterAI (OpenAI-compatible API)
openai==2.17.0

# Data processing
pydantic==2.5.3
PyMuPDF==1.26.7
pillow==12.1.0
